# APISpec/generated/models/user-entity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * The main API for all Sandbox apps
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import { AddressEntity } from "./address-entity";

/**
 *
 * @export
 * @interface UserEntity
 */
export interface UserEntity {
	/**
	 *
	 * @type {string}
	 * @memberof UserEntity
	 */
	id?: string;
	/**
	 *
	 * @type {string}
	 * @memberof UserEntity
	 */
	firstName: string;
	/**
	 *
	 * @type {string}
	 * @memberof UserEntity
	 */
	lastName: string;
	/**
	 *
	 * @type {string}
	 * @memberof UserEntity
	 */
	email: string;
	/**
	 *
	 * @type {string}
	 * @memberof UserEntity
	 */
	phone?: string;
	/**
	 *
	 * @type {number}
	 * @memberof UserEntity
	 */
	birthDate?: number;
	/**
	 *
	 * @type {string}
	 * @memberof UserEntity
	 */
	gender?: string;
	/**
	 *
	 * @type {AddressEntity}
	 * @memberof UserEntity
	 */
	address?: AddressEntity;
}