# APISpec/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/users/list":{"post":{"operationId":"listUsers","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiPaginatedRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseListEntity"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}]}}}}},"tags":["users"]}},"/users/{userId}":{"get":{"operationId":"getUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"]}},"/accounts":{"get":{"operationId":"getAccounts","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"API","description":"The main API for all Sandbox apps","version":"1.0","contact":{}},"tags":[{"name":"users","description":""},{"name":"accounts","description":""}],"servers":[],"components":{"schemas":{"ResponseListEntity":{"type":"object","properties":{"data":{"type":"array","items":{"type":"string"}},"total":{"type":"number"}},"required":["data"]},"AddressEntity":{"type":"object","properties":{"id":{"type":"number"},"lineOne":{"type":"string"},"lineTwo":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zipCode":{"type":"string"}},"required":["lineOne","city","state","zipCode"]},"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"birthDate":{"type":"number"},"gender":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressEntity"}},"required":["firstName","lastName","email"]},"EnumFilterType":{"type":"string","enum":["Contains","GreaterThan","Search"]},"FilterType":{"type":"object","properties":{"value":{"oneOf":[{"type":"string"},{"type":"number"},{"type":"boolean"}]},"type":{"$ref":"#/components/schemas/EnumFilterType"},"field":{"type":"string"}},"required":["value","type"]},"ApiPaginatedRequest":{"type":"object","properties":{"start":{"type":"number"},"limit":{"type":"number"},"page":{"type":"number"},"filters":{"type":"array","items":{"$ref":"#/components/schemas/FilterType"}}},"required":["start","limit","page"]}}}}